[matchingImageCollection] split functions in cpp
